Reorganize work pools card into subdirectory
Move dashboard work pools card files into work-pools-card subdirectory for better organization. Update imports and add index.ts for cleaner exports.
